BUG 18277082 - FTS: STACK BUFFER OVERFLOW IN INNOBASE_STRNXFRM AND TIS620
When we run windows debug build or use a linux Address Sanitizer build,
there would be a failure or stack-buffer-overflow in innobase_strnxfrm.

The root cause is that my_strnxfrm_tis620 would access one more byte of
the dst parameter and set it to '\0', which is unnecessary in this case.
So we just do not set this terminating '\0'.

Approved by Tor, rb#5051.