Problem:
In the clustered index, when an update operation is done the overall
scenario is as follows:
          
1.  Delete mark the old record that is to be updated.
2.  The old record disowns the blobs. As part of this step, we assert that
    the blobs are owned by the old record.
3.  Checks the foreign key constraints.
4.  Insert the new record into clustered index.
          
Scenario involving DB_LOCK_WAIT:
          
If step 3 times out, then we will repeat steps 1 and 2.  Since the steps 1
and 2 are already done, the assert that the blobs must be owned in step 2
will fail.

Solution:
          
If step 3 times out, then skip steps 1 and 2 and proceed with 3.  This is
achieved by checking if the record is already delete marked.

rb#5081 approved by Marko.