Bug#18064775 Assertion failed: fields.elements == values.elements Bug#11745889 Traditional: INSERT accepts invalid date from default value
A recent bugfix (16820562) gave column count zero a special meaning,
but failed to give an error for the case of zero column names but more
than zero column values.

The fix is to check explicitly for this case. The interface for
check_insert_fields() has been extended with a bool value_count_known,
to make it easy to remove it when the "known" property is no longer
needed (all values are actually "known", but the code path is clumsy).

The bugfix also re-introduced bug 11745889, which was fixed by bug
16078943, alas it did not contain a test for this particular case.

The fix for this bug is to treat INSERT statements with no column names
and no column values specified as if a partial (but empty) column list
has been specified. This way, the write_set for the table is not
populated and the columns will be picked for validation of default
values.