Bug#18338203: P_S REPLICATION_CONNECTION_STATUS.SOURCE_UUID               UNINITIALIZED VALUE
Problem:
The master_uuid field was introduced in mysql-5.6.0.                            
When replicating from a pre-5.6.0 master to a 5.7 slave,                        
the slave will ask for the master_uuid field which the                          
master doesnt have. Prior to fix, the P_S code made an                          
assuption that if there is a master there will be a                             
master_uuid for the master which is not true in the case                        
of say 5.5->5.7 replication. Since the P_S code                                 
blindly copied MASTER_UUID bytes from mi object,                                
we ended up with a garbage value in the select                                  
SOURCE_UUID field from P_S.replication_connect_status 

Fix:
- Added a check for empty master_uuid. In this case
  we print an empty value.