Bug #16102788: INDETERMINATE BEHAVIOR DUE TO EMPTY OPTION VALUES
Currently mysql programs silently accept empty strings as option values,
e.g. "--port=" will work the same as "--port=0" and -P "" will work the same as
-P 0.
This is unintuitive and can cause a lot of confusion.
Fixed the options collection code to throw an error if the explicit argument to 
an option is empty string and the option is not handled programatically and 
does have an option value pointer to be set.

Test case added.