Bug#18304086 - P_S DIGEST LOOKS WRONG FOR SYSTEM VARIABLES, SHOWN AS @ @ VARIABLE
Before this fix, a query containing "@@variable",
when printing the query digest text,
was printed as "@ @ variable".

This is a bug, because the syntax "@ @ variable" is actually not accepted
by the server.

"@@variable" is parsed as 3 tokens, '@', '@' and 'variable'.
When printing the digest text, a space is added to separate each token,
causing the resulting string.

The root cause of the bug is in the lexer:
if '@' was a real token, "@ @ variable" would also be legal,
and the server would also accept the "@ @ variable" syntax,
so returning '@' as a token alone is incorrect.

Now, independently of the lexer, query digest text printed could use
a better formatting in general, so the token description has been expanded
to have an append_space flag, so that a space is not systematically added
for each token.

This fix does not change the lexer in any way.
This fix only prints the '@' without a trailing space,
in the query digest text.