Bug#18259356 - PROCESSLIST_ID FOR SLAVE SQL THREAD IS MISSING
For historical reasons, background replication threads are
assigned a connection_id(), and displayed in:
- SHOW PROCESSLIST
- table INFORMATION_SCHEMA.PROCESSLIST.

In the performance schema, these threads are also displayed,
but with a NULL PROCESSLIST_ID.

Having the same thread displayed:
- with a PROCESSLIST_ID (in the information_schema)
- without a PROCESSLIST_ID (in the performance_schema)
is inconsistent, and is a bug.

With this fix, the ID assigned in the server code
is also given to the instrumentation,
so that column PROCESSLIST_ID for replication threads
is correctly populated in table performance_schema.threads.