Bug #17457755: MYISAM FTS: CRASH IN FT_SIMPLE_GET_WORD
The UPDATE command was trying to prepare (by calling 
setup_ftfuncs() ) the full text index search functions before 
calling fix_fields() on the update values.
This obviously is not a good idea, specially if the fulltext 
search is part of the update value.

Moved the call to setup_ftfuncs() after the preparation
of the UPDATE values. 

Test case added.