WL#7806: Detect the current system tablespace size.
We still skip the test if the tablespace name differs. The reason for this
is that the test would hang if --mysqld=--innodb-data-file-path is passed.
This is because the parameter would override whatever is specified in the
configuration file. Also, we do not escape any ; that might have been
specified in --innodb-data-file-path.