Remove an invalid sentence from a comment that was introduced in WL#7142.
During a log checkpoint, there may exist dirty blocks in the buffer pool.
What matters is that before the checkpoint completes, we will have
flushed all blocks up to the checkpoint LSN, and no blocks with a newer LSN.

In the event of a crash, we will recover the system as it was at the
checkpoint LSN, after applying any redo log that has been
successfully written since the latest checkpoint.