BUG#18484649 - STRICT MODE + CURSOR FETCH INTO WRONG TYPE,                ASSERTION IN PROTOCOL::END_STATEMENT
BACKGROUND:
An assertion failure in 5.7 in Protocol::end_statement()
occurred when we try to fetch a large value (int) which
cannot fit into a variable (tiny int) declared inside
stored routine through a cursor. However, it works fine
in 5.6.
Also, This assertion failure happens only in debug builds,
We do not get any assertion failure/crash on release builds,
Because asserts are not active on release builds.

ANALYSIS:
On analysis, it is found that error comes in
Materialized_cursor::fetch() function. But its return type
is void. Therefore, the calls to upper level functions assumed
everything went OK with it (however, it resulted in error),
so it resulted in incorrect value of err_status in
sp_head::execute() function.
The documentation for
virtual bool execute(THD *thd, uint *nextp) = 0;
function in sp_instr.h says

    @return Error status.

In this case execute() did not perform as documented - it has
reported an error but still returned false.

FIX:
The fix is to set err_status variable correctly (to true if
it is known that some error has occurred or the query was
killed i.e thd->killed is set).
The return type of function Materialized_cursor::fetch() is
modified from void to bool, so that it reports the error to
upper level functions correctly. This fix results in err_status
variable inside sp_head::execute() to be set correctly
after the execution of statement and perform according to
the documentation.