Bug#18556403 USE STD LIBRARY FOR SORTING AND MATCHING IN SUBCLASSES OF IN_VECTOR, PART 2
As demonstrated by the patch for Bug#18486249
in_vector and its derived classes can benefit from some code modernization.

standard library functions are
 - type safe
 - easier to read/maintain
 - faster

Introduce typed vectors of objects, rather than the "blob vector" in the base class in_vector.
The re-write uncovered a bug in the range optimizer:
it was building range predicates for unused parts of the IN-vector.