ndb_index_stat_partition.test failed on some platforms after increasing number of LDM threads to 4. This was due to index statistics only being sampled from one of the LDM's. Thus increasing #LDM's, also decreased the size of the sample, which gave more variations in the statistics.
As ndb_index_stat_partition.test required the estimated number of rows
to within 45% of the actuall number of rows, more variations in the 
statistics caused this test to fail.

This fix doubles the number of rows in the test table in
order to give better statistcs estimates when number of 
LDMs are increased.