Bug#18561062: MYSQL-TRUNK REVNO 7736 CAUSED A -2% ~ -3%               PERFORMANCE REGRESSION (OLTP_RO)
The regression is caused by calculating the condition filtering
effect for simple single table queries where condition filtering
has no effect.

Whether or not condition filtering shall be calculated is based
on a number of conditions. One of these conditions is whether
join buffering may be used for the table.

Optimize_table_order::calculate_scan_cost() called 
calculate_condition_filter() with a wrong parameter that indicated
that join buffering was used even when that was not the case.

Fixed by passing "join_buffering_is_used" parameter instead of
unconditional 'true'.