Bug#18599982: HAVE_DECL_FDATASYNC & HAVE_FEDISABLEEXCEPT               CMAKE CONFIGURE CHECKS ARE BROKEN
The problem was that the results of the HAVE_DECL_FDATASYNC and
HAVE_FEDISABLEEXCEPT CMake configure checks were not stored
in my_config.h due to them being absent from the config.h.cmake
template.

This meant that code relying on these checks was in fact broken.
The affected code was used to prefer fdatasync() over fsync()
and to disable floating point exceptions.

The patch fixes the problem by updating config.h.cmake so
that the results of these two checks are saved.

The patch also reorganizes config.h.cmake to make it easier to
match the file with configure.cmake and other CMake files where
checks are defined. In addition, the patch removes some duplicate
CMake checks and some unused items from config.h.cmake.