Removed the MTR test ndb_mt_recv.test:
This test was introduced in order to have a minimum of testing
with multiple receiver threads in the ThreadConfig. The testload
itself for this test, was the ndb_basic.test

As multiple receiver threads now has been made part of default_ndbd.cnf,
this special variant of ndb_basic.test is obsolete.