ndb_backup_rate.test checks that we are able to recover normal operation after a 'Redo log full' situation.
This is done by configuring a slow 'DiskCheckpointSpeed=1M', and
letting the test program produce lots of logging by inserting and
deleting rows while waiting for a 'Redo log full' error.

However, when running with debug compiled binaries, the insert + delete
performance could be so slow that we are unable to fill the redo log.
Thus this test often failed in the 'mix-debug' part of the test.

This fix disable the test when run with debug binaries.
NOTE: It was already disabled for Valgrind for the same reasons.