Bug#18596756 - FAILED PREPARING OF TRIGGER ON TRUNCATED TABLES                CAUSE ERROR 1054.
Analysis:
--------
Issue here is, the re-parse of the 'SELECT' query in the 'SET'
statement for the NEW field of a row (triggered due to a DDL
operation), removes current object and creates a new object for
the "NEW.<field>" information. But the binding between the new
object and the actual field in the table is not set. Hence while
setting the value, an "unknown column" error is reported.

For the SET operation with NEW.<field>, object of type
"sp_instr_set_trigger_field" is created while parsing the trigger.
This object has the member "m_trigger_field" of type
"Item_trigger_field"(Represents NEW/OLD field of row). This
trigger is bound with actual field of the row by calling method
"Item_trigger_field::fix_field".

During the cleanup process, performed before the re-parse of the 
"SELECT" query of "SET" operation, "Item_trigger_field" 
field held by "m_trigger_field" is unlinked. After the "SELECT"
query is re-parsed, the "Item_trigger_field" for "NEW.id" is created
and held with "m_trigger_field" member of "sp_instr_set_trigger_field"
object. But it is not bound with the actual field object of table. 
Hence an "unknown column" is reported.  

Fix:
---
Modified code to bind the "Item_trigger_field" created for "NEW.id"
to the Field object in table by calling "Item_trigger_field::fix_field"
method.