This is a patch for Bug#18068253 - XA START WITH THE EXIST XID, AND THEN XA COMMIT WITH THE XID WILL FAILED.
When a xa transaction is prepared and the server crashed, then after restart the 
MySQL server,  first, do XA START with the same xid, says:" XAER_DUPID: The 
XID already exists", that's OK, but do XA COMMIT with this xid, says that:
XAER_RMFAIL: The command cannot be executed when global transaction is 
in the  NON-EXISTING state. So, the prepared transaction can't be committed. 

The reason for the bug was that the method eq() of class xid didn't take into
account the value of xid::format during comparison. First time when XA  START
is executed the server found prepared transaction and return error but for all
of that the value of thd->transaction.xid_state.xid is reset assigning the
value -1 to the the attribute xid.format. Later when the statement XA COMMIT
is executed comparison of  thd->transaction.xid_state.xid with xid value
specified in the statement is made without taking into account
the value of format.