Bug#17882305 "#SQL-*" TABLES SHOULD BE SKIPPED BY NDB_RESTORE
Added an option --exclude-intermediate-tables, default TRUE, to 
ndb_restore. When this is set to TRUE, ndb_restore skips restoring 
tables whose names are prefixed with "#sql", thus skipping intermediate 
tables created during ALTER TABLE operations.

Added a unit test to autotest to create an NDB table with a 
"#sql"-prefixed name and take a backup. This backup is then restored 
to a database which does not contain the "#sql"-prefixed table.