Bug #17449901	 TABLE DISAPPEARS WHEN ALTERING WITH 			FOREIGN KEY CHECKS OFF
Analysis:
By removing the FK index with out removing the foreign key constraint,
we are moving the table into inconsistent state. FK index lookup happens
during load table process. So it fails to open the table when FK index
is missing.

Fix:
Don't allow to drop an index if it is needed for an FK constraint,
even if foreign_key_checks is disabled.

		Approved by Jimmy (rb-5274)