Bug#18330694: INNODB TESTCASES RUN TOO SLOW ON WINDOWS
This fix contains 3 different parts

1) Reason for mysqltest command shutdown server always waited the full
timeout:
During shutdown of the server there are still handles open to the process
(by mtr?), 
resulting in that it is possible to retrieve a new handle to the process,
which leads to the current my_kill(pid, 0) windows implementation thinks
the process is still alive, when it infact already exited.

So it needs to either check with WaitForSingleObject or GetExitCodeProcess
to verify if the process handle really has exited or not.

2) Reason for spurious test failures when doing copy_file/remove_file/
move_file:
Even if the process signaled that it exited, Windows may not yet closed
all the process files, resulting in failures (ERROR_SHARING_VIOLATION)
during move/delete/copy. So to be more tolerant I added on retry in such
cases.

3) Killing the server if the shutdown takes too long time.
Changing shutdown_server to:
If timeout is non-zero, send shutdown to the server and wait until the
server terminates or timeout is exceeded. If timeout is exceeded return
error.
If timeout is zero, then kill/terminate the server (kill -9 on Unix,
TerminateProcess on Windows).
This since killing the server if timeout is exceeded can have side-effects
that is hard to predict. Perhaps we should never kill in shutdown_server
(not allowing zero as timeout) and introduce kill_server
(current shutdown_server 0 behavior)? I have not done this since it
would need to update several test files.

I also included a test change that I used during testing.