Bug#18503515: UNION with set function in ORDER BY should be rejected
A UNION query is not an aggregated query, and an ORDER BY applied to it
cannot contain aggregate functions. This patch prevents such queries.

A new error message is added: ER_AGGREGATE_ORDER_FOR_UNION.

It was also found that nest_level is not correctly updated for
fake_select_lex. This has been remedied, but to our knowledge, there
were no error from this mistake.