BUG#17798076 - BUG#16418661 CLEANUP: REMOVE UNNECESSARY BUF_FLUSH_LIST()                DURING RECOVERY
Why and when the extra buf_flush_list() was added?
--------------------------------------------------
The extra buf_flush_list() was added as part of rb#2293.
The testcase for rb#2293 failed randomly reporting inconsistent number
of undo tablespaces. MTR didn't provide a way to use custom
innodb_data_file_path, innodb_undo_tablespaces. See Bug#17059436 So
the workaround is to use "--exec $MYSQLD --args my.cnf" on a empty data
directory. This will create  necessary files and aborts after that.

The extra flush made sure that correct the number of undo tablespaces
reported. This is not correct because the changes are already covered by redo.

So the the ideal solution would be have been to run "--exec $MYSQLD on
the datadir" again to apply redo or allow the bootstrapping to finish
properly and exit.

Fix:
----
Remove uncessary flush. Fix the inconsistent number of undo tablespaces
issue by passing '--skip-grant-tables --innodb-unkown-parameter' to $MYSQLD.
This would allow the bootstrapping to finish and properly exit.

Approved by Marko, Kevin. rb#5234, rb#5320