Bug#18675099: WINDOWSCACHE.CMAKE MUST BE UPDATED FOR VS2013
WindowsCache.cmake hardcodes results for CMake configure checks so they
don't have to be performed on Windows (as this is very slow).

The problem was that the cache file had several values which were no
longer correct. The consequence that we were using Windows specific
workarounds in cases where it was no longer needed.

This patch removes the outdated values from the cache file so that the
checks are actually made. It also reorganizes the cache file so that it
matches config.h.cmake, removes duplicate entries, removes dead
entries and adds missing entries. The patch also removes some
unneeded Visual Studio version checks.