BUG#18642372 - I_INNODB.INNODB-LOG-FILE THROW TIMEOUT ERROR ONLY ON WINDOWS
Problem:

The reason for hang is because DeleteFile() hangs on ibtmp1. 
The file is opened by the page cleaner thread. The handle is not in
Datafile Object but it is stored in fil_node_t.

The explicit close() on Datafile will not solve the issue because the
handle is not opened by Datafile Object. It is stored in fil_node_t by page
cleaner thread.

Fix:
Exit flush thread early incase of startup failure. This will make sure
that we don't have handle open on 'ibtmp1' by the page cleaner thread.
Without any open handles, deletion of ibtmp1 will be successful

Note:
Although it->close() in Tablespace::delete_files() will not solve the
current hang problem, it will make sure that we close the file before deleting.

Approved by Kevin,Yasufumi,Krunal rb#5258