Bug#11762867 - DEFAULT VALUES FOR INITLOGFILEGROUP CAUSE ERROR 2350: 'INVALID CONFIGURAT
Problem:

The config.ini specifies that a log file group should be created
during initial start of the cluster with the help of
InitialLogFileGroupSpec. When no undo_buffer_size=<size> is specified,
the default value for undo_buffer_size will be used.

 When cluster starts initial with default values, it fails to allocate
 undo buffer of the requested size from the SharedGlobalMemory
 pool. This is very likely in 7.0-1, where defaults for SGA is 20M and
 undo_buffer_size is 64M.

 Suggested solution:

1) At initial startup, allocate the memory for the undo_buffer_size
specified in the InitialLogFileGroupSpec or the default, in addition
to the sga.

2)Improve the error message printed when ndbd detects that it
cannot allocate the undo buffer.