Bug#18363515 MULTIPLE THREADED SLAVE MAY BE CRASHED BY              STOP SLAVE SQL
Worker error will crash MTS if coordinator is waiting for a
free worker to dispatch next event. The crash happens because
the caller of Mts_submode_logical_clock::get_least_occupied_worker()
supposes it always return a valid worker. But that is not true.
It may turn NULL when coordinator is aborted at the same time.

This patch makes the caller of get_least_occupied_worker() to
check if the function returns a valid value.