Bug #18723872  FK: ASSERTION: *CURSOR->INDEX->NAME == TEMP_INDEX_PREFIX
This is a regression caused by the fix for Bug #11758237 INSERT ON DUPLICATE
KEY UPDATE SOMETIMES WRITES BINLOG POSITION INCORRECT.  The trx->error_state
will be modified when a thread is suspended.  But logic of placing just the gap
locks on unique secondary indexes without actually performing the insert
depends on the value of trx->error_state.  So the value of trx->error_state is
restored for insert into every index.  Refer to rb#3196 for the original fix.

approved by Marko over IM.