Bug#11758766:MYSQLD CONTINUES OPERATION WITHOUT LOGGING WHEN BINLOGS CANNOT BE WRITTEN
Problem:
========
If an error occurs that prevents mysqld writing to the
binary logs (disk full, readonly file system, etc) then the
logs are disabled and operations continue. This can lead to
out of sync slaves and improper backups.

Analysis:
=========
When binlogging becomes impossible due to readonly file
system at present the binlog gets closed along with an error
message in error log which says binlog is turned off for the
whole duration of the server to fix the cause shutdown the
server and fix the problem. The master continues without
binlogging which causes the slave to go out of sync. At
present there is no way user can control the behaviour so it
will be better to let the user provide an option to stop the
server immediately on the error.

Fix:
====
A new option "binlogging_impossible_mode" has been
introduced whose values are "IGNORE" or "ABORT". When
binlogging becomes impossible if user sets the variable to
"ABORT" server will stop if user sets it to "IGNORE" binlog
will be turned off and server will continue.